LayoutTests/platform/glib/imported/w3c/web-platform-tests/cookies/value/value-ctl-expected.txt


PASS Cookie with %x0 in value is truncated.
FAIL Cookie with %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=10" but got "test=10; value"
PASS Cookie with %xd in value is truncated.
FAIL Cookie with %x1 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=1"
FAIL Cookie with %x2 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=2"
FAIL Cookie with %x3 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=3"
FAIL Cookie with %x4 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=4"
FAIL Cookie with %x5 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=5"
FAIL Cookie with %x6 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=6"
FAIL Cookie with %x7 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=7"
FAIL Cookie with %x8 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=8"
FAIL Cookie with %x9 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=9"
FAIL Cookie with %xb in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=11"
FAIL Cookie with %xc in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=12"
FAIL Cookie with %xe in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=14"
FAIL Cookie with %xf in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=15"
FAIL Cookie with %x10 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=16"
FAIL Cookie with %x11 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=17"
FAIL Cookie with %x12 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=18"
FAIL Cookie with %x13 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=19"
FAIL Cookie with %x14 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=20"
FAIL Cookie with %x15 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=21"
FAIL Cookie with %x16 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=22"
FAIL Cookie with %x17 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=23"
FAIL Cookie with %x18 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=24"
FAIL Cookie with %x19 in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=25"
FAIL Cookie with %x1a in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=26"
FAIL Cookie with %x1b in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=27"
FAIL Cookie with %x1c in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=28"
FAIL Cookie with %x1d in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=29"
FAIL Cookie with %x1e in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=30"
FAIL Cookie with %x1f in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=31"
FAIL Cookie with %x7f in value is rejected. assert_equals: The cookie was rejected. expected "" but got "test=127value"
PASS Cookie with %x1 after %x0 in value is truncated.
PASS Cookie with %x2 after %x0 in value is truncated.
PASS Cookie with %x3 after %x0 in value is truncated.
PASS Cookie with %x4 after %x0 in value is truncated.
PASS Cookie with %x5 after %x0 in value is truncated.
PASS Cookie with %x6 after %x0 in value is truncated.
PASS Cookie with %x7 after %x0 in value is truncated.
PASS Cookie with %x8 after %x0 in value is truncated.
PASS Cookie with %x9 after %x0 in value is truncated.
PASS Cookie with %xb after %x0 in value is truncated.
PASS Cookie with %xc after %x0 in value is truncated.
PASS Cookie with %xe after %x0 in value is truncated.
PASS Cookie with %xf after %x0 in value is truncated.
PASS Cookie with %x10 after %x0 in value is truncated.
PASS Cookie with %x11 after %x0 in value is truncated.
PASS Cookie with %x12 after %x0 in value is truncated.
PASS Cookie with %x13 after %x0 in value is truncated.
PASS Cookie with %x14 after %x0 in value is truncated.
PASS Cookie with %x15 after %x0 in value is truncated.
PASS Cookie with %x16 after %x0 in value is truncated.
PASS Cookie with %x17 after %x0 in value is truncated.
PASS Cookie with %x18 after %x0 in value is truncated.
PASS Cookie with %x19 after %x0 in value is truncated.
PASS Cookie with %x1a after %x0 in value is truncated.
PASS Cookie with %x1b after %x0 in value is truncated.
PASS Cookie with %x1c after %x0 in value is truncated.
PASS Cookie with %x1d after %x0 in value is truncated.
PASS Cookie with %x1e after %x0 in value is truncated.
PASS Cookie with %x1f after %x0 in value is truncated.
PASS Cookie with %x7f after %x0 in value is truncated.
FAIL Cookie with %x1 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=1term" but got "test=1term; va"
FAIL Cookie with %x2 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=2term" but got "test=2term; va"
FAIL Cookie with %x3 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=3term" but got "test=3term; va"
FAIL Cookie with %x4 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=4term" but got "test=4term; va"
FAIL Cookie with %x5 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=5term" but got "test=5term; va"
FAIL Cookie with %x6 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=6term" but got "test=6term; va"
FAIL Cookie with %x7 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=7term" but got "test=7term; va"
FAIL Cookie with %x8 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=8term" but got "test=8term; va"
FAIL Cookie with %x9 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=9term" but got "test=9term; va"
FAIL Cookie with %xb after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=11term" but got "test=11term; va"
FAIL Cookie with %xc after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=12term" but got "test=12term; va"
FAIL Cookie with %xe after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=14term" but got "test=14term; va"
FAIL Cookie with %xf after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=15term" but got "test=15term; va"
FAIL Cookie with %x10 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=16term" but got "test=16term; va"
FAIL Cookie with %x11 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=17term" but got "test=17term; va"
FAIL Cookie with %x12 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=18term" but got "test=18term; va"
FAIL Cookie with %x13 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=19term" but got "test=19term; va"
FAIL Cookie with %x14 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=20term" but got "test=20term; va"
FAIL Cookie with %x15 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=21term" but got "test=21term; va"
FAIL Cookie with %x16 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=22term" but got "test=22term; va"
FAIL Cookie with %x17 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=23term" but got "test=23term; va"
FAIL Cookie with %x18 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=24term" but got "test=24term; va"
FAIL Cookie with %x19 after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=25term" but got "test=25term; va"
FAIL Cookie with %x1a after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=26term" but got "test=26term; va"
FAIL Cookie with %x1b after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=27term" but got "test=27term; va"
FAIL Cookie with %x1c after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=28term" but got "test=28term; va"
FAIL Cookie with %x1d after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=29term" but got "test=29term; va"
FAIL Cookie with %x1e after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=30term" but got "test=30term; va"
FAIL Cookie with %x1f after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=31term" but got "test=31term; va"
FAIL Cookie with %x7f after %xa in value is truncated. assert_equals: The cookie was set as expected. expected "test=127term" but got "test=127term; value"
PASS Cookie with %x1 after %xd in value is truncated.
PASS Cookie with %x2 after %xd in value is truncated.
PASS Cookie with %x3 after %xd in value is truncated.
PASS Cookie with %x4 after %xd in value is truncated.
PASS Cookie with %x5 after %xd in value is truncated.
PASS Cookie with %x6 after %xd in value is truncated.
PASS Cookie with %x7 after %xd in value is truncated.
PASS Cookie with %x8 after %xd in value is truncated.
PASS Cookie with %x9 after %xd in value is truncated.
PASS Cookie with %xb after %xd in value is truncated.
PASS Cookie with %xc after %xd in value is truncated.
PASS Cookie with %xe after %xd in value is truncated.
PASS Cookie with %xf after %xd in value is truncated.
PASS Cookie with %x10 after %xd in value is truncated.
PASS Cookie with %x11 after %xd in value is truncated.
PASS Cookie with %x12 after %xd in value is truncated.
PASS Cookie with %x13 after %xd in value is truncated.
PASS Cookie with %x14 after %xd in value is truncated.
PASS Cookie with %x15 after %xd in value is truncated.
PASS Cookie with %x16 after %xd in value is truncated.
PASS Cookie with %x17 after %xd in value is truncated.
PASS Cookie with %x18 after %xd in value is truncated.
PASS Cookie with %x19 after %xd in value is truncated.
PASS Cookie with %x1a after %xd in value is truncated.
PASS Cookie with %x1b after %xd in value is truncated.
PASS Cookie with %x1c after %xd in value is truncated.
PASS Cookie with %x1d after %xd in value is truncated.
PASS Cookie with %x1e after %xd in value is truncated.
PASS Cookie with %x1f after %xd in value is truncated.
PASS Cookie with %x7f after %xd in value is truncated.
